parser_rss/pyproject.toml
[tool.poetry]
name = "parser-rss"
version = "0.1.0"
description = ""
authors = ["alex <me@fomin3.ru>"]
readme = "README.md"
packages = [{include = "parser_rss"}]

[tool.poetry.dependencies]
python = "^3.10"
asyncio = "^3.4.3"
aiohttp = {extras = ["speedups"], version = "^3.8.3"}
loguru = "^0.6.0"
aioredis = "^2.0.1"
pydantic = "^1.10.2"
feedparser = "^6.0.10"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
aio-pika = "^8.2.5"
aerich = "^0.7.1"



[tool.aerich]
tortoise_orm = "parser_rss.config.POSTGRES_CONFIG_AERICH"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
